Fix problem with exact images

+++ src/+replab/+rep/RepByImages_exact.m
@@ -83,7 +83,7 @@ classdef RepByImages_exact < replab.RepByImages
                 % group structure to perform computations
                 target1 = replab.GeneralLinearGroupWithInverses(self.field, self.dimension, true);
                 target2 = replab.GeneralLinearGroup(self.field, self.dimension, true);
-                cut = replab.Morphism.lambda(target1, target2, @(X) double(X(:, 1:self.dimension)));
+                cut = replab.Morphism.lambda(target1, target2, @(X) X(:, 1:self.dimension));
                 images = arrayfun(@(i) [images{i} inverseImages{i}], 1:m, 'uniform', 0);
                 res = replab.bsgs.ChainWithImages.make(n, target1, nicePreimages, images, base, order);
                 res = res.mapImages(cut);
